requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
boto3==1.18.0
botocore==1.21.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.51.0
django-cors-headers==3.11.0
django-filter==21.1
django-hosts==5.1
django-jquery==3.1.0
django-pytest==0.2.0
django-rest==0.8.7
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
iniconfig==1.1.1
jmespath==0.10.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.0.0
pytest-django==4.5.2
python-dateutil==2.7.3
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.5.0
six==1.16.0
smmap==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.10
uWSGI==2.0.20
whitenoise==5.2.0
